Module pos_ledger_db::schema

This module defines representation of Diem core data structures at physical level via schemas that implement schemadb::schema::Schema.

All schemas are pub(crate) so not shown in rustdoc, refer to the source code to see details.

Constants

BLOCK_BY_EPOCH_AND_ROUND_CF_NAME

COMMITTED_BLOCK_BY_VIEW_CF_NAME

COMMITTED_BLOCK_CF_NAME

EPOCH_BY_VERSION_CF_NAME

EVENT_ACCUMULATOR_CF_NAME

EVENT_BY_KEY_CF_NAME

EVENT_BY_VERSION_CF_NAME

EVENT_CF_NAME

JELLYFISH_MERKLE_NODE_CF_NAME

LEDGER_COUNTERS_CF_NAME

LEDGER_INFO_BY_BLOCK_CF_NAME

LEDGER_INFO_BY_VOTED_BLOCK_CF_NAME

POS_STATE_CF_NAME

REWARD_EVENT_CF_NAME

STALE_NODE_INDEX_CF_NAME

TRANSACTION_ACCUMULATOR_CF_NAME

TRANSACTION_BY_ACCOUNT_CF_NAME

TRANSACTION_CF_NAME

TRANSACTION_INFO_CF_NAME