pos_ledger_db

Struct PosLedgerDB

pub struct PosLedgerDB { /* private fields */ }

This holds a handle to the underlying DB responsible for physical storage and provides APIs for access to the core Diem data structures.

Implementations

impl PosLedgerDB

pub fn open<P: AsRef<Path> + Clone>( db_root_path: P, readonly: bool, prune_window: Option<u64>, rocksdb_config: RocksdbConfig, ) -> Result<Self>

pub fn update_rocksdb_properties(&self) -> Result<()>

This force the db to update rocksdb properties immediately.

pub fn get_transaction(&self, version: Version) -> Result<Transaction>

pub fn get_transaction_block_meta( &self, version: Version, ) -> Result<Option<(Version, BlockMetadata)>>

pub fn get_transaction_info(&self, version: u64) -> Result<TransactionInfo>

pub fn get_backup_handler(&self) -> BackupHandler

Gets an instance of BackupHandler for data backup purpose.